Use `tangotango` theme
This configuration isn’t the best, in that it unconditionally sets
the theme to `tangotango`, even in the terminal, when I wouldn’t want
it to.  I need to modify this at some point, but since I usually use
Emacs through a GUI, this isn’t my most pressing need.

License: GPLv3+