Remove `flyspell-correct`
This is part of the general cleanup for removing Ivy.  I used to use
this a lot more, but I’ve been fine with just `ispell` and
`flyspell-auto-correct-word` (bound to `M-TAB`) recently.  While this
does mean I don’t have the easy interface for adding words to my
dictionary, I still have `ispell` for that.

License: GPLv3+