Remove Ivy configuration
Vertico has been sufficiently solid for me that I don’t think I will
need to suddenly disable it and re-enable Ivy, so I’m removing my Ivy
configuration.  If I do need it in the future, it’s still in the
curated git history (a nice byproduct of making my configuration
public), so I can retrieve it again later.

At this point, I haven’t replaced Counsel and Swiper with Consult yet,
or Hydra with Embark.  With the former, I will need to play around
with it a lot more to be comfortable, but I am already finding I miss
Hydras in certain circumstances.  I’ve tried to use Swiper a few
times, consciously needing to swap to ~M-x occur~, which works well
enough.  Perhaps that will be the bigger push for me to try Consult.

License: GPLv3+