Remove `use-package` hook suffix
By default `use-package` will tack on the suffix `-hook` to any hook
you configure.  While this the normal case, there are some
pathological hooks that are not so properly named.  By turning this
off, we trade off the additional noise of more `-hook`s for the
ability to configure these poorly-named hooks.

License: GPLv3+