Remove `org-roam-ui` as a submodule
This is left over from when `org-roam-ui` was not on MELPA, so I had
to include it manually in my repository.  To preserve the `lisp/`
subdirectory, though, which I use sometimes for the infrequent
packages that are not in a package repository, we add an empty file to
make sure it still exists.

License: GPLv3+