Install `diminish` package if it is not installed
Now that `use-package` is included in Emacs, our old check of
installing both it and `diminish` if `use-package` is not installed no
longer works—`use-package` may still be installed even if `diminish`
is no longer installed.  This patch changes the logic to try to
install both packages if either is not installed.

License: GPLv3+