Use git repo for `typo`
The name for this package was silently taken oven by a different
package, which has caused me a few issues in the past.  However, MELPA
has always had the package that I want.

However, as of [This commit][melpa-commit], MELPA has removed the
package I’ve been using, in favor of the one who stole the name, which
is also packaged in GNU ELPA.  It looks like this put the maintaner
tarsius between a rock and a hard place.  I don’t blame him.  But this
means I have to vendor in the package, or download it from git in
Emacs.

This patch teaches our `use-package` declaration to download the
package we want from its canonical GitHub repository.

[melpa-commit]: https://github.com/melpa/melpa/commit/11e080fa5a065ca4c2a3eff1203f80d7d1b1c597

License: GPLv3+